requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
einops==0.7.0
filelock==3.13.1
fsspec==2024.2.0
ftfy==6.2.3
huggingface-hub==0.24.5
idna==3.4
importlib_metadata==8.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
packaging==24.1
pillow==10.2.0
psutil==6.0.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.28.1
safetensors==0.4.4
sentencepiece==0.1.97
sympy==1.12
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.0
triton==2.1.0
typing_extensions==4.9.0
urllib3==1.26.13
wcwidth==0.2.13
xformer==1.0.1
zipp==3.19.2
datasets==2.21.0
accelerate==0.33.0
diffusers==0.28.0
bitsandbytes
tensorboard
peft==0.12.0
opencv-python
pyrallis